HW4/HW4_Flask/requirements.txt
Flask==2.1.3
numpy==1.23.1
Pillow==9.2.0
requests==2.22.0
